requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==1.11.18
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
urllib3==1.25.3
